Rename redux slice + testing changes in AccountUpdate.test
